chore: utilsディレクトリをapplication配下に移動
