+++ src/application/commands/implementations/auth.ts
@@ -10,8 +10,8 @@ import Department from "../../../domain/entities/department";
 import type Member from "../../../domain/entities/member";
 import type Command from "../../../domain/types/command";
 import { adminAuth } from "../../../infrastructure/firebase";
-import addRoleToMember from "../../../utils/addRoleToMember";
-import createRoleIfNotFound from "../../../utils/createRoleNotFound";
+import addRoleToMember from "../../utils/addRoleToMember";
+import createRoleIfNotFound from "../../utils/createRoleNotFound";
 import {
   getMemberByDiscordIdController,
   getMemberByEmailController,
