+++ src/application/commands/implementations/hotChannel.ts
@@ -1,6 +1,6 @@
 import { type CommandInteraction, SlashCommandBuilder } from "discord.js";
 import type Command from "../../../domain/types/command";
-import checkIsAdmin from "../../../utils/checkMemberRole";
+import checkIsAdmin from "../../utils/checkMemberRole";
 import { generateChannelActivityRanking } from "../../usecases/getHotChannels";
 
 const hotChannelsCommand: Command = {
