+++ src/application/commands/implementations/kill.ts
@@ -1,6 +1,6 @@
 import { type CommandInteraction, SlashCommandBuilder } from "discord.js";
 import type Command from "../../../domain/types/command";
-import checkIsAdmin from "../../../utils/checkMemberRole";
+import checkIsAdmin from "../../utils/checkMemberRole";
 
 const killCommand: Command = {
   data: new SlashCommandBuilder()
