+++ src/application/commands/implementations/register.ts
@@ -1,7 +1,7 @@
 import { type CommandInteraction, SlashCommandBuilder } from "discord.js";
 import Department from "../../../domain/entities/department";
 import type Command from "../../../domain/types/command";
-import checkIsAdmin from "../../../utils/checkMemberRole";
+import checkIsAdmin from "../../utils/checkMemberRole";
 import { addMemberController } from "../../controllers/MemberController";
 import type { MemberCreateInput } from "../../repository/IMemberRepository";
