+++ src/application/commands/implementations/renameAll.ts
@@ -1,7 +1,7 @@
 import { type CommandInteraction, SlashCommandBuilder } from "discord.js";
 import type Command from "../../../domain/types/command";
 import logger from "../../../infrastructure/logger";
-import checkIsAdmin from "../../../utils/checkMemberRole";
+import checkIsAdmin from "../../utils/checkMemberRole";
 import { getMemberByDiscordIdController } from "../../controllers/MemberController";
 
 const renameALL: Command = {
