--- src/utils/clearAuthData.ts
@@ -1,13 +1,0 @@
-import type AuthData from "../domain/types/authData";
-
-function clearAuthData(): AuthData {
-  return {
-    name: undefined,
-    student_number: undefined,
-    department: undefined,
-    mail: undefined,
-    discordId: undefined,
-  };
-}
-
-export default clearAuthData;
